+++ numpy/matrixlib/tests/test_masked_matrix.py
@@ -7,7 +7,7 @@ from numpy.ma.core import (masked_array, masked_values, masked, allequal,
                            MaskType, getmask, MaskedArray, nomask,
                            log, add, hypot, divide)
 from numpy.ma.extras import mr_
-from numpy.core.numeric import pickle
+from numpy.compat import pickle
 
 
 class MMatrix(MaskedArray, np.matrix,):
